feat: Pages RouterからApp Routerへの移行とnext-sitemapからNext.js標準機能への移行
- Pages RouterからApp Routerに移行
- next-sitemapからNext.js標準のsitemap.xml/robots.txt生成機能に移行
- SSRを無効化してSSGのみで動作するように設定
- Material-UIコンポーネントをApp Router対応に修正
- 依存関係のクリーンアップ（next-sitemapパッケージ削除）
