+++ .gitignore
@@ -1,6 +1,7 @@
 node_modules/*
 
 .next/
+dist/
 
 robots.txt
 sitemap*
