--- next-sitemap.config.js
@@ -1,9 +1,0 @@
-const config = {
-  siteUrl: process.env.NEXT_PUBLIC_DOMAIN_URL || 'http://localhost:3000',
-  changefreq: 'monthly',
-  generateRobotsTxt: true,
-  sitemapSize: 7000,
-  outDir: './public',
-}
-
-export default config
