+++ package.json
@@ -3,11 +3,10 @@
   "private": true,
   "scripts": {
     "dev": "next",
-    "build": "next build",
+    "build": "next build && next export",
     "start": "next start",
     "format": "prettier -w ./src",
     "prebuild": "eslint .",
-    "postbuild": "next-sitemap",
     "ogp": "node --loader ts-node/esm ./bin/ogImage/create.mts"
   },
   "dependencies": {
@@ -35,7 +34,6 @@
     "eslint": "9.37.0",
     "eslint-config-flat-gitignore": "^2.1.0",
     "eslint-config-next": "15.5.4",
-    "next-sitemap": "^4.2.3",
     "prettier": "^3.6.2",
     "ts-node": "^10.9.2",
     "typescript": "^5.9.3"
