+++ src/components/PageViewTracker.tsx
@@ -1,0 +1,8 @@
+'use client'
+
+import usePageView from '../hooks/usePageView'
+
+export default function PageViewTracker() {
+  usePageView()
+  return null
+}
