+++ src/components/uiParts/Footer/index.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import Link from 'next/link'
 import React from 'react'
 import Grid from '@mui/material/Grid'
