+++ src/components/uiParts/Header/index.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import * as React from 'react'
 import Image from 'next/image'
 import Link from 'next/link'
