+++ src/components/uiParts/JapanMap/index.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import Link from 'next/link'
 
 import cssStyle from './style.module.scss'
