+++ src/components/uiParts/Layout/index.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import * as React from 'react'
 import Box from '@mui/material/Box'
 import MuiContainer from '@mui/material/Container'
