+++ src/components/uiParts/Text/index.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import { TypographyProps } from '@mui/material'
 import Typography from '@mui/material/Typography'
