+++ src/hooks/usePageView.ts
@@ -1,22 +1,14 @@
+'use client'
+
 import { useEffect } from 'react'
-import { useRouter } from 'next/router'
+import { usePathname } from 'next/navigation'
 
 import { pageview } from '../lib/gtag'
 
 export default function usePageView() {
-  const router = useRouter()
+  const pathname = usePathname()
 
   useEffect(() => {
-    const handleRouteChange = (path: string) => {
-      pageview(path)
-    }
-
-    router.events.on('hashChangeComplete', handleRouteChange)
-    router.events.on('routeChangeComplete', handleRouteChange)
-
-    return () => {
-      router.events.off('hashChangeComplete', handleRouteChange)
-      router.events.off('routeChangeComplete', handleRouteChange)
-    }
-  }, [router.events])
+    pageview(pathname)
+  }, [pathname])
 }
